requirements.txt
# Flask and web framework
Flask==2.3.2
Werkzeug==2.3.6
Jinja2==3.1.2

# Environment variables
python-dotenv==1.0.0

# OpenAI API
openai==1.3.0

# Document processing
PyPDF2==3.0.1
pdfkit==1.0.0
reportlab==4.0.4
python-docx==1.0.1

# Security and authentication
cryptography==41.0.3

# Data handling
jsonschema==4.19.0
requests==2.31.0
uuid==1.30

# Datetime handling
pytz==2023.3

# File operations
python-magic==0.4.27

# Web server for deployment
gunicorn==21.2.0

# Image processing
Pillow==10.0.0

# Natural language processing
langdetect==1.0.9

# Regular expressions and text processing
regex==2023.8.8